Got a wasm compiled version of the core reschedule module mostly working
It doesn't work in SSR, and there are aspects of the integration that are not great at all, but it does work!
Going to try to refine it down a bit, and consider how much I care about SSR. It's not strictly important for this actual project, but I'd like to see it working on principle.