Sped up the boundary crossing code by making it more manual/low level
I'm still not super happy with this, but it's not actually much less typesafe then the json/serde based version, and it's substantially faster, if slightly more fiddly. I'd work more on it if this was going to be the first of many more in depth integrations, but it's a one off goofy project, so I'm gonna move on.